pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-anamnesis"
version = "0.1.0"
description = "A dbt-like platform for building and managing flashcard projects"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Open Anamnesis Contributors"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=5.4.0",
    "jsonschema>=4.0.0",
    "jinja2>=3.0.0",
    "flask>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
]

[project.scripts]
anamnesis = "open_anamnesis.cli:main"
anamnesis-compile = "open_anamnesis.cli:compile_cmd"
anamnesis-build = "open_anamnesis.cli:build_cmd"

[project.urls]
Homepage = "https://github.com/open-anamnesis/open-anamnesis"
Repository = "https://github.com/open-anamnesis/open-anamnesis.git"
Issues = "https://github.com/open-anamnesis/open-anamnesis/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
